services/backend/requirements.txt
Click==7.0
fastapi==0.65.2
h11==0.9.0
pydantic==1.6.2
#redis==3.3.11
redis>=3.4.0
rq==1.2.0
starlette
uvicorn==0.11.7
pandas==2.0.1
pytz==2023.3
PyJWT==2.8.0
aiomysql==0.2.0
aioredis==2.0.1
uvloop
cryptography
yfinance==0.2.18
celery